refactor: use workspace and crates structure
