+++ crates/bwq/src/commands/mod.rs
@@ -1,0 +1,7 @@
+pub mod check;
+pub mod examples;
+pub mod server;
+
+pub use check::run_check;
+pub use examples::run_examples;
+pub use server::run_server;
