+++ crates/bwq_linter/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bwq_linter"
+version = "0.0.0"
+authors = { workspace = true }
+edition = { workspace = true }
+description = "Core linting library for Brandwatch query files (.bwq)"
+repository = { workspace = true }
+license = { workspace = true }
+publish = false
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+pretty_assertions = { workspace = true }
+test-case = { workspace = true }
