+++ crates/bwq_linter/resources/test/fixtures/invalid/invalid_mixed_operators.bwq
