+++ crates/bwq_linter/resources/test/fixtures/valid/comments_and_wildcards.bwq
