+++ crates/bwq_linter/resources/test/fixtures/valid/complex_boolean_operations.bwq
