+++ crates/bwq_linter/resources/test/fixtures/valid/complex_field_combinations.bwq
