+++ crates/bwq_linter/resources/test/fixtures/valid/complex_near.bwq
