+++ crates/bwq_linter/resources/test/fixtures/valid/complex_proximity_operations.bwq
