+++ crates/bwq_linter/resources/test/fixtures/valid/complex_social_media.bwq
