+++ crates/bwq_linter/resources/test/fixtures/valid/field_operations.bwq
