+++ crates/bwq_linter/resources/test/fixtures/valid/valid_multiline.bwq
