+++ crates/bwq_linter/src/error.rs
