+++ crates/bwq_linter/src/lexer.rs
