+++ crates/bwq_linter/src/validation/rules/operator_rules.rs
