+++ crates/bwq_linter/src/validation/rules/performance_rules.rs
