+++ crates/bwq_linter/src/validator.rs
