+++ crates/bwq_server/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "bwq_server"
+version = "0.0.0"
+authors = { workspace = true }
+edition = { workspace = true }
+description = "Language server for Brandwatch query files (.bwq)"
+repository = { workspace = true }
+license = { workspace = true }
+publish = false
+
+[dependencies]
+anyhow = { workspace = true }
+bwq_linter = { workspace = true }
+crossbeam-channel = { workspace = true }
+lsp-server = { workspace = true }
+lsp-types = { workspace = true }
+serde_json = { workspace = true }
+tokio = { workspace = true }
