+++ crates/bwq_server/src/utils.rs
@@ -1,6 +1,6 @@
 use lsp_types::{Position as LspPosition, Range as LspRange};
 
-use crate::error::{Position, Span};
+use bwq_linter::error::{Position, Span};
 
 pub fn position_to_lsp(pos: &Position) -> LspPosition {
     LspPosition {
