--- src/lsp/mod.rs
@@ -1,5 +1,0 @@
-pub mod diagnostics_handler;
-pub mod server;
-pub mod utils;
-
-pub use server::LspServer;
