--- src/main.rs
@@ -1,369 +1,0 @@
-use std::fs;
-use std::path::{Path, PathBuf};
-
-use clap::{Parser, Subcommand};
-use ignore::WalkBuilder;
-use rayon::prelude::*;
-
-use bwq::analyze_query;
-
-#[derive(Parser)]
-#[command(name = "bwq")]
-#[command(about = "A linter for Brandwatch query files (.bwq)")]
-#[command(version = "0.2.0")]
-struct Cli {
-    #[command(subcommand)]
-    command: Option<Commands>,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-    /// lint files, directories, or query strings
-    #[command(name = "check")]
-    Check {
-        /// Files or directories to check (ignored if --query is used) [default: .]
-        files: Vec<PathBuf>,
-
-        /// Lint a query string directly (instead of files)
-        #[arg(long, short = 'q')]
-        query: Option<String>,
-
-        /// Suppress warning messages
-        #[arg(long)]
-        no_warnings: bool,
-
-        /// Output format (text or json)
-        #[arg(long, default_value = "text")]
-        output_format: String,
-
-        /// Exit with status code 0, even upon detecting lint violations
-        #[arg(long)]
-        exit_zero: bool,
-
-        /// File extensions to check (can be used multiple times)
-        #[arg(long = "extension", short = 'e', default_values = ["bwq"])]
-        extensions: Vec<String>,
-    },
-
-    /// Show example queries
-    Examples,
-
-    /// Start LSP server
-    Lsp,
-}
-
-fn main() {
-    let cli = Cli::parse();
-
-    match cli.command {
-        Some(Commands::Check {
-            files,
-            query,
-            no_warnings,
-            output_format,
-            extensions,
-            exit_zero,
-        }) => {
-            if let Some(query_str) = query {
-                lint_single_query_string(&query_str, !no_warnings, &output_format, exit_zero);
-            } else if files.is_empty() {
-                // Default to current directory when no files specified
-                let default_files = vec![PathBuf::from(".")];
-                process_files(
-                    &default_files,
-                    !no_warnings,
-                    &output_format,
-                    &extensions,
-                    exit_zero,
-                );
-            } else {
-                process_files(&files, !no_warnings, &output_format, &extensions, exit_zero);
-            }
-        }
-
-        Some(Commands::Examples) => {
-            show_examples();
-        }
-        Some(Commands::Lsp) => {
-            if let Err(e) = bwq::lsp::LspServer::run() {
-                eprintln!("LSP server error: {e}");
-                std::process::exit(1);
-            }
-        }
-        None => {
-            eprintln!("Error: A subcommand is required");
-            eprintln!("\nUsage: bwq <COMMAND>");
-            eprintln!("\nCommands:");
-            eprintln!("  check        Lint files, directories, or queries");
-            eprintln!("  examples     Show example queries");
-            eprintln!("  lsp          Start LSP server");
-            eprintln!("\nFor more information, try 'bwq --help'");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn process_files(
-    files: &[PathBuf],
-    show_warnings: bool,
-    output_format: &str,
-    extensions: &[String],
-    exit_zero: bool,
-) {
-    // Validate that all paths exist
-    for file_path in files {
-        if !file_path.exists() {
-            eprintln!("Path does not exist: {}", file_path.display());
-            if !exit_zero {
-                std::process::exit(1);
-            }
-            return;
-        }
-    }
-
-    // Use unified linting approach
-    if lint_paths(files, show_warnings, output_format, extensions).is_err() && !exit_zero {
-        std::process::exit(1);
-    }
-}
-
-fn lint_single_query_string(
-    query: &str,
-    show_warnings: bool,
-    output_format: &str,
-    exit_zero: bool,
-) {
-    let analysis = analyze_query(query);
-
-    match output_format {
-        "json" => {
-            output_json(&analysis);
-        }
-        _ => {
-            output_text(&analysis, show_warnings);
-        }
-    }
-
-    if !analysis.is_valid && !exit_zero {
-        std::process::exit(1);
-    }
-}
-
-fn matches_extensions(file_path: &Path, extensions: &[String]) -> bool {
-    if let Some(file_ext) = file_path.extension().and_then(|ext| ext.to_str()) {
-        extensions.iter().any(|ext| ext == file_ext)
-    } else {
-        false
-    }
-}
-
-fn discover_files(paths: &[PathBuf], extensions: &[String]) -> Vec<PathBuf> {
-    let mut discovered_files = Vec::new();
-
-    for path in paths {
-        if path.is_file() {
-            // For explicit file arguments, include them regardless of extension
-            discovered_files.push(path.clone());
-        } else if path.is_dir() {
-            // For directories, discover files with matching extensions
-            let mut builder = WalkBuilder::new(path);
-            builder.hidden(false);
-
-            for dir_entry in builder.build().flatten() {
-                let file_path = dir_entry.path();
-                if file_path.is_file() && matches_extensions(file_path, extensions) {
-                    discovered_files.push(file_path.to_path_buf());
-                }
-            }
-        }
-    }
-
-    discovered_files
-}
-
-fn lint_paths(
-    paths: &[PathBuf],
-    show_warnings: bool,
-    output_format: &str,
-    extensions: &[String],
-) -> Result<(), ()> {
-    let files = discover_files(paths, extensions);
-
-    if files.is_empty() {
-        eprintln!(
-            "No files found that have the extension(s): {}",
-            extensions.join(", ")
-        );
-        return Err(());
-    }
-
-    // process files in parallel
-    let results: Vec<_> = files
-        .par_iter()
-        .filter_map(|file_path| {
-            let content = match fs::read_to_string(file_path) {
-                Ok(content) => content,
-                Err(e) => {
-                    eprintln!("Error reading file {}: {}", file_path.display(), e);
-                    return None;
-                }
-            };
-
-            let query = content.trim();
-            let analysis = analyze_query(query);
-            Some((file_path.clone(), analysis, query.to_string()))
-        })
-        .collect();
-
-    let total_files = results.len();
-    let valid_files = results
-        .iter()
-        .filter(|(_, analysis, _)| analysis.is_valid)
-        .count();
-    let any_errors = results.iter().any(|(_, analysis, _)| !analysis.is_valid);
-
-    match output_format {
-        "json" => {
-            let mut errors = Vec::new();
-            let mut warnings = Vec::new();
-
-            for (file_path, analysis, _) in &results {
-                for error in &analysis.errors {
-                    let mut error_json = error.to_json();
-                    if let Some(obj) = error_json.as_object_mut() {
-                        obj.insert(
-                            "filename".to_string(),
-                            serde_json::Value::String(file_path.display().to_string()),
-                        );
-                    }
-                    errors.push(error_json);
-                }
-
-                for warning in &analysis.warnings {
-                    let mut warning_json = warning.to_json();
-                    if let Some(obj) = warning_json.as_object_mut() {
-                        obj.insert(
-                            "filename".to_string(),
-                            serde_json::Value::String(file_path.display().to_string()),
-                        );
-                    }
-                    warnings.push(warning_json);
-                }
-            }
-
-            let output = serde_json::json!({
-                "summary": {
-                    "total_files": total_files,
-                    "valid_files": valid_files,
-                    "invalid_files": total_files - valid_files
-                },
-                "errors": errors,
-                "warnings": warnings
-            });
-
-            println!("{}", serde_json::to_string_pretty(&output).unwrap());
-        }
-        _ => {
-            for (file_path, analysis, _) in &results {
-                if !analysis.is_valid || (show_warnings && !analysis.warnings.is_empty()) {
-                    println!("File: {}", file_path.display());
-                    output_text(analysis, show_warnings);
-                    println!();
-                }
-            }
-
-            println!("Summary: {valid_files}/{total_files} files valid");
-        }
-    }
-
-    if any_errors { Err(()) } else { Ok(()) }
-}
-
-fn output_text(analysis: &bwq::AnalysisResult, show_warnings: bool) {
-    println!("{}", analysis.summary());
-
-    if !analysis.errors.is_empty() {
-        println!("\nErrors:");
-        for (i, error) in analysis.errors.iter().enumerate() {
-            println!("  {}. {}: {}", i + 1, error.code(), error);
-        }
-    }
-
-    if show_warnings && !analysis.warnings.is_empty() {
-        println!("\nWarnings:");
-        for (i, warning) in analysis.warnings.iter().enumerate() {
-            println!("  {}. {}: {}", i + 1, warning.code(), warning);
-        }
-    }
-}
-
-fn output_json(analysis: &bwq::AnalysisResult) {
-    let errors: Vec<_> = analysis.errors.iter().map(|e| e.to_json()).collect();
-    let warnings: Vec<_> = analysis.warnings.iter().map(|w| w.to_json()).collect();
-
-    let json_output = serde_json::json!({
-        "query": analysis.query,
-        "errors": errors,
-        "warnings": warnings
-    });
-
-    println!("{}", serde_json::to_string_pretty(&json_output).unwrap());
-}
-
-fn show_examples() {
-    println!("Brandwatch Query Examples:");
-    println!();
-
-    println!("Basic Boolean Operators:");
-    println!("  apple AND juice");
-    println!("  apple OR orange");
-    println!("  apple NOT bitter");
-    println!("  (apple OR orange) AND juice");
-    println!();
-
-    println!("Quoted Phrases:");
-    println!("  \"apple juice\"");
-    println!("  \"organic fruit\" AND healthy");
-    println!();
-
-    println!("Proximity Operators:");
-    println!("  \"apple juice\"~5");
-    println!("  apple NEAR/3 juice");
-    println!("  apple NEAR/2f juice");
-    println!();
-
-    println!("Wildcards and Replacement:");
-    println!("  appl*");
-    println!("  customi?e");
-    println!();
-
-    println!("Field Operators:");
-    println!("  title:\"apple juice\"");
-    println!("  site:twitter.com");
-    println!("  author:brandwatch");
-    println!("  language:en");
-    println!("  rating:[3 TO 5]");
-    println!();
-
-    println!("Location Operators:");
-    println!("  country:usa");
-    println!("  region:usa.ca");
-    println!("  city:\"usa.ca.san francisco\"");
-    println!();
-
-    println!("Advanced Operators:");
-    println!("  authorFollowers:[1000 TO 50000]");
-    println!("  engagementType:RETWEET");
-    println!("  authorGender:F");
-    println!("  {{BrandWatch}}  (case-sensitive)");
-    println!();
-
-    println!("Comments:");
-    println!("  apple <<<This is a comment>>> AND juice");
-    println!();
-
-    println!("Special Characters:");
-    println!("  #MondayMotivation");
-    println!("  @brandwatch");
-    println!();
-}
